+++ src/k8s-configuration/azext_k8s_configuration/providers/FluxConfigurationProvider.py
@@ -46,19 +46,19 @@ from ..validators import (
     validate_url_with_params,
 )
 from .. import consts
-from ..vendored_sdks.v2022_01_01_preview.models import (
+from ..vendored_sdks.v2022_03_01.models import (
     FluxConfiguration,
     FluxConfigurationPatch,
     GitRepositoryDefinition,
     GitRepositoryPatchDefinition,
     BucketDefinition,
+    BucketPatchDefinition,
     RepositoryRefDefinition,
     KustomizationDefinition,
     KustomizationPatchDefinition,
-    DependsOnDefinition,
     SourceKindType,
 )
-from ..vendored_sdks.v2021_09_01.models import Extension, Identity
+from ..vendored_sdks.v2022_03_01.models import Extension, Identity
 
 logger = get_logger(__name__)
 
@@ -384,17 +384,10 @@ def create_kustomization(
             consts.CREATE_KUSTOMIZATION_EXIST_HELP,
         )
 
-    # Add the dependencies in their model to the kustomization
-    model_dependencies = None
-    if dependencies:
-        model_dependencies = []
-        for dep in parse_dependencies(dependencies):
-            model_dependencies.append(DependsOnDefinition(kustomization_name=dep))
-
     kustomization = {
         kustomization_name: KustomizationPatchDefinition(
             path=path,
-            depends_on=model_dependencies,
+            depends_on=parse_dependencies(dependencies),
             timeout_in_seconds=parse_duration(timeout),
             sync_interval_in_seconds=parse_duration(sync_interval),
             retry_interval_in_seconds=parse_duration(retry_interval),
@@ -451,17 +444,10 @@ def update_kustomization(
             consts.UPDATE_KUSTOMIZATION_NO_EXIST_HELP,
         )
 
-    # Add the dependencies in their model to the kustomization
-    model_dependencies = None
-    if dependencies:
-        model_dependencies = []
-        for dep in parse_dependencies(dependencies):
-            model_dependencies.append(DependsOnDefinition(kustomization_name=dep))
-
     kustomization = {
         kustomization_name: KustomizationPatchDefinition(
             path=path,
-            depends_on=model_dependencies,
+            depends_on=parse_dependencies(dependencies),
             timeout_in_seconds=parse_duration(timeout),
             sync_interval_in_seconds=parse_duration(sync_interval),
             retry_interval_in_seconds=parse_duration(retry_interval),
@@ -1015,7 +1001,7 @@ class BucketGenerator(SourceKindGenerator):
 
         def bucket_patch_updater(config):
             if any(kwarg is not None for kwarg in self.kwargs.values()):
-                config.bucket = BucketDefinition(
+                config.bucket = BucketPatchDefinition(
                     url=self.url,
                     bucket_name=self.bucket_name,
                     timeout_in_seconds=parse_duration(self.timeout),