+++ src/k8s-configuration/HISTORY.rst
@@ -3,6 +3,11 @@
 Release History
 ===============
 
+1.5.0
+++++++++++++++++++
+* Update models to 2022-03-01 for GA
+* Remove unneeded warning for HTTPS urls
+
 1.4.1
 ++++++++++++++++++
 * [BREAKING CHANGE] `--access-key` changed to `--bucket-access-key`