+++ src/spring-cloud/HISTORY.md
@@ -1,8 +1,12 @@
 Release History
 ===============
-3.0.2
+3.1.1
 ---
 * Revert new RBAC requirement for Standard and Basic sku Spring resource for `az spring-cloud app set-deployment` and `az spring-cloud app unset-deployment` commands.
+
+3.1.0
+---
+* Add support for user-assigned managed identity on App (Preview).
 
 3.0.1
 ---